Comments and formatting
Updated comments into every file and added an enum for Receipt. Prevents
Receipt from entering a state outside of the 3 specified. I also added
some new style for loop stuff for D. Vogs, but it isn't working yet. It
is commented out for now so the code will run as expected.